FWF-4277: [Feature] Rename anonymous_application.py to public_endpoints.py
