+++ forms-flow-api/src/formsflow_api/resources/__init__.py
@@ -11,7 +11,6 @@ from formsflow_api_utils.exceptions import (
 )
 from formsflow_api_utils.utils.constants import ALLOW_ALL_ORIGINS
 
-from formsflow_api.resources.anonymous_application import API as PUBLIC_API
 from formsflow_api.resources.application import API as APPLICATION_API
 from formsflow_api.resources.application_history import (
     API as APPLICATION_HISTORY_API,
@@ -29,6 +28,7 @@ from formsflow_api.resources.import_support import API as IMPORT_API
 from formsflow_api.resources.ipaas import API as INTEGRATION_API
 from formsflow_api.resources.metrics import API as APPLICATION_METRICS_API
 from formsflow_api.resources.process import API as PROCESS_API
+from formsflow_api.resources.public_endpoints import API as PUBLIC_API
 from formsflow_api.resources.roles import API as KEYCLOAK_ROLES_API
 from formsflow_api.resources.theme import API as THEME_CUSTOMIZATION_API
 from formsflow_api.resources.user import API as KEYCLOAK_USER_API
